\name{mynewpackage-package}
\alias{mynewpackage-package}
\alias{mynewpackage}
\docType{package}
\title{
\packageTitle{mynewpackage}
}
\description{
\packageDescription{mynewpackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mynewpackage}
\packageIndices{mynewpackage}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{mynewpackage}

Maintainer: \packageMaintainer{mynewpackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
